[FIX] UserRating tests + GenreAndRating tests

+++ recommender-persistence/src/test/java/cz/muni/fi/pa165/genre/GenreAndRatingTest.java
@@ -25,7 +25,7 @@ public class GenreAndRatingTest {
 
         this.correctSmall = new GenreAndRating(genre,0f);
         this.correctBig = new GenreAndRating(genre,10f);
-        this.correctNull = new GenreAndRating(genre,null);
+        this.correctNull = new GenreAndRating(genre,(Float) null);
     }
 
     @Test
